requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
alembic==1.5.8
antlr4-python3-runtime==4.8
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autokeras==1.0.12
av==8.0.3
Babel==2.9.0
backcall==0.2.0
bayesian-bootstrap==1.0.5
bleach==3.3.0
blessings==1.7
cachetools==4.2.1
catboost==0.25.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cliff==3.7.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorful==0.5.4
colorlog==4.8.0
configparser==5.0.2
ConfigSpace==0.4.18
coolname==1.1.0
cycler==0.10.0
Cython==0.29.22
daemonize==2.5.0
dagging @ git+https://github.com/chkoar/dagging.git@cac364f0f40e2b5144068ba44855949eda6d7643
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
docker-pycreds==0.4.0
entrypoints==0.3
ffmpeg-python==0.2.0
filelock==3.0.12
FLAML==0.3.0
flatbuffers==1.12
fsspec==0.8.7
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.14
google-api-core==1.26.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
graphviz==0.16
greenlet==1.0.0
grpcio==1.32.0
guildai==0.7.3
h5py==2.10.0
hiredis==2.0.0
hpbandster==0.7.4
hydra-core==1.0.6
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
ipykernel==5.5.3
ipympl==0.7.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keras-tuner==1.0.2
kiwisolver==1.3.1
lightgbm==3.2.1
lightly==1.1.4
llvmlite==0.36.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
marshmallow-enum==1.5.1
matplotlib==3.3.4
mistune==0.8.4
mpld3==0.5.2
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.9
networkx==2.5.1
notebook==6.3.0
numba==0.53.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
omegaconf==2.0.6
opencensus==0.7.12
opencensus-context==0.1.2
opt-einsum==3.3.0
optuna==2.7.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
patsy==0.5.1
pbr==5.5.1
pickle5==0.0.11
pickleshare==0.7.5
Pillow==8.1.2
pkginfo==1.7.0
plotly==4.14.3
prettytable==2.1.0
probatus==1.7.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
py-spy==0.3.5
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyperclip==1.8.2
pyreadline==2.1
Pyro4==4.80
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-lightning==1.0.4
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
ray==1.2.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scikit-optimize==0.8.1
scipy==1.5.3
seaborn==0.11.1
Send2Trash==1.5.0
sentry-sdk==1.0.0
serpent==1.30.2
shap==0.39.0
shortuuid==1.0.1
SimpleITK==2.0.2
six==1.15.0
slicer==0.0.7
smmap==4.0.0
sniffio==1.2.0
SQLAlchemy==1.4.7
statsmodels==0.12.2
stevedore==3.3.0
subprocess32==3.5.4
tabulate==0.8.9
tabview==1.4.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
terminaltables==3.1.0
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.31
torch==1.8.1+cu111
torchaudio==0.8.1
torchmetrics==0.2.0
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
tune-sklearn==0.2.1
typeguard==2.11.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
videoio==0.1.2
virtualenv==20.4.3
wandb==0.10.24
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
windows-curses==2.2.0
wrapt==1.12.1
xgboost==1.3.3
yarl==1.6.3
zipp==3.4.1
